scb_trials/requirements.txt
# API Clients
openai>=1.0.0
google-genai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0
numpy>=1.24.0

# PDF processing
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Azure services
azure-ai-formrecognizer>=3.3.0
azure-identity>=1.15.0

# Google Cloud services
google-cloud-documentai>=2.20.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
